[chore]: 미사용 코드 삭제 후후 스토리북  추가
